pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "alpaca-mcp-server"
version = "2.0.0"
description = "Alpaca Trading API integration for Model Context Protocol (MCP)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

authors = [
    {name = "Alpaca", email = "info@alpaca.markets"}
]

keywords = ["mcp", "alpaca", "trading", "finance", "ai", "llm", "model-context-protocol"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]

dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.urls]
Homepage = "https://alpaca.markets/"
Repository = "https://github.com/alpacahq/alpaca-mcp-server"
"Bug Tracker" = "https://github.com/alpacahq/alpaca-mcp-server/issues"
Documentation = "https://github.com/alpacahq/alpaca-mcp-server#readme"

[project.scripts]
alpaca-mcp-server = "alpaca_mcp_server.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/alpaca_mcp_server"]

[tool.hatch.build]
include = [
    "src/alpaca_mcp_server/**/*.py",
    "src/alpaca_mcp_server/specs/*.json",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = ["-v", "--tb=short"]
